CAL-329 Moved alliance-branding-plugin from distribution to platform
