+++ distribution/pom.xml
@@ -28,7 +28,6 @@
     <name>Alliance :: Distribution</name>
 
     <modules>
-        <module>alliance-branding-plugin</module>
         <module>sdk</module>
         <module>common</module>
         <module>install-profiles</module>
