+++ platform/alliance-branding-plugin/src/main/java/org/codice/alliance/branding/impl/AllianceBrandingPlugin.java
@@ -11,7 +11,7 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.alliance.distribution.branding;
+package org.codice.alliance.branding.impl;
 
 import java.io.IOException;
 import java.io.InputStream;
