+++ platform/alliance-branding-plugin/src/main/resources/META-INF/branding.properties
