+++ platform/alliance-branding-plugin/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -14,7 +14,7 @@
 <blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0">
 
     <service interface="org.codice.ddf.branding.BrandingPlugin">
-        <bean class="org.codice.alliance.distribution.branding.AllianceBrandingPlugin" init-method="init">
+        <bean class="org.codice.alliance.branding.impl.AllianceBrandingPlugin" init-method="init">
             <argument value="/META-INF/branding.properties"/>
         </bean>
     </service>
