+++ platform/alliance-branding-plugin/src/main/resources/alliance/alliance-logo.png
