+++ platform/alliance-branding-plugin/src/main/resources/alliance/favicon.ico
