+++ platform/alliance-branding-plugin/src/main/resources/alliance/webconsole.css
