+++ platform/alliance-branding-plugin/src/test/java/org/codice/alliance/branding/impl/AllianceBrandingPluginTest.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
 
-package org.codice.alliance.distribution.branding;
+package org.codice.alliance.branding.impl;
 
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
